Speed and error handling improvements; log requests
Several places now have extra logic to catch and print exceptions to the
console.

On a successful request, basic information about the request and how
long it took is logged to a postgres table so it can be reported by
minerva.

When creating an epub, the language is now uniformly set to English
(rather than mostly English with a little Croatian).

The epub creator now tries to request all fic chapters together, but
falls back to the old behavior of requesting one chapter per request
with a limit of 64 active threads. Redundant fic info requests are no
longer made, the metadata already fetched is passed in.

The filename of the generated epub is now more safely constructed and
includes the Alexandria id as a disambiguator.