es: very basic elasticsearch functionality
Whenever fic metadata is looked up we attempt to save it in
elasticsearch as well.

When an export fails, we do a _very_ basic lookup in elasticsearch to
offer suggestions.

Since the returned fixits may be much longer and contain html some
changes have been made to the /legacy_epub_export template and the json
error pretty printer for js-enabled clients.

The es file can also be run directly to reindex data.