db.ExportLog: new; track and reuse previous exports
This has two major benefits now and lays groundwork for further
improvements. First, for converted formats like mobi and pdf it is much
faster and less resource intensive if the actual conversion process can
be skipped. Secondly, for those same formats, by skipping the conversion
when we can we sidestep the issue of their non-reproducibility.

This does not include any changes to the way epubs are generated --
likely we would have upstream track the meta and content hashes and base
our determination on that. For all other formats, we use an internal
counter (essentially a cache buster) and the epub hash as the key. This
means that _every_ export still generates an epub first.