/fic: use generate links if file is out of date
This fixes the case where a secondary format was generated some time ago
and then the epub was updated. Previously, we'd simply generate a direct
download link to the last known version of each individual format so the
secondary file would be out of date and the user would need to go
through the re-export workflow.

Now, if the secondary format is out of date we build a generate link
instead. This fixes the non-js workflow where re-export can only be used
to update the epub.