db.FicBlacklist: new; prevent a fic from being exported or showing up
This introduces a new blacklist feature to prevent a fic from being
exported or from existing information or exports from appearing. This
combined with manually deleting the cached exports and calling the
es.blacklist method should cover fully removing a fic at the behest of
an author.

Further cleanup and ensuring there are no edge cases is needed.